Fixed HealthController issue that was causing test failures.
It turns out that my previous change meant to address issues with the
HealthController was incomplete (I didn't cover all code paths), and it
cause the timestamps to be incorrect.
(380516eb4c204b44ced7b894d52cce6737229fa9)

This fix fetches all of the samples in the specified window, ensures
they are chronologically ordered and returns only the most recent.