Added a template for creation of a container registry secret.
Previously it was necessary to pre-create a secret with registry
credentials in the namespace that this helm char was being installed in.
This additional template allows the credentials to be provided as part
of the values and creates the secret as part of the chart. This mimics
the behavior of BigBang charts.