Refactored the sonar-agent Helm Chart's service config values.
Service configuration is now specified with dicts instead of arrays for
configs, services, and healthChecks.

This allows configmaps that are used as sources for these values to be
merged with better semantics. In short, if a configmap for a particular
environment needs change just one property of one particular health
check it can specify just that one property, instead of having to
replicate all of the rest of the configuration. Health check conditions
would still have to be replaced all or nothing because they are not
uniquely identified by some key.

Also made some tweaks to the way some appsettings values were layed out
and the config checksum generation.