Fixed issue with SONAR Agent generated API client
Returning a collection of tuples from an API controller results in
generated code that expects a customed named type to exist to represent
the Tuple, but no such type exists in our model classes. As a result the
generated SONAR Agent client does not compile.

This change creates a custom type to wrap the collection of tuples
returned by the SONAR API. This pattern should be followed for future
API endpoints.