Refactored HealthCheckList and RootService/ChildService components
There were some pathological API queries being run in the
HealthCheckListItem component because data wasn't being queries and
passed to child components in an appropriate way.

Specifically each HealthCheckListItem needed the HealthCheckModel so
that it could display the health check definition. Each
HealthCheckListItem instance was fetching the complete configuration
hierarchy for the current tenant, and then selecting any health checks
with matching name. In addition to excess data queries this resulted in
the multiple health checks being displayed because health check names
are not globally unique within a service hierarchy.