Historical health check data (UI and new endpoint)
Closes BATAPI-426

## Description:

Adds an endpoint that returns historical health check results at a given timestamp for a service. I chose to add an endpoint instead of appending this data to the ServiceHierarchyHealthHistory model to help performance, as we really only need to query this data when the status history drawer is opened by a user/a different Status History Tile is selected. Makes corresponding changes in the UI.

## Security Impact Analysis Questionnaire

### Submitter Checklist
-  [ ] Is there an impact on Auditing and Logging procedures or capabilities?
-  [ ] Is there an impact on Authentication procedures or capabilities?
-  [ ] Is there an impact on Authorization procedures or capabilities?
-  [ ] Is there an impact on Communication Security procedures or capabilities?
-  [ ] Is there an impact on Cryptography procedures or capabilities?
-  [ ] Is there an impact on Sensitive Data procedures or capabilities?
-  [ ] Is there an impact on any other security-related procedures or capabilities?
-  [x] No security impacts identified.

## Security Risks Identified:

<!-- For any applicable items on the "Submitter Checklist," describe the impact of the change and any implemented mitigations. -->

Closes BATAPI-426